+++ test/integrated_tests/core/authc/test_authentication.py
@@ -1,9 +1,10 @@
 import pytest
 from yosai.core import (
+    AdditionalAuthenticationRequired,
     AuthenticationException,
-    Credential,
     InvalidAuthenticationSequenceException,
     SerializationManager,
+    SimpleIdentifierCollection,
     UsernamePasswordToken,
     AuthenticationException,
 )
@@ -11,23 +12,23 @@ from yosai.core import (
 
 def test_single_factor_authc_userpass_using_accountstore_success(
         caplog, default_authenticator, valid_username_password_token,
-        event_bus):
+        event_bus, cache_handler):
     da = default_authenticator
     event_detected = None
 
-    def event_listener(identifiers=None):
+    def event_listener(identifier=None):
         nonlocal event_detected
-        event_detected = identifiers
+        event_detected = identifier
     event_bus.register(event_listener, 'AUTHENTICATION.SUCCEEDED')
 
-    account = da.authenticate_account(None, valid_username_password_token)
+    account_id = da.authenticate_account(None, valid_username_password_token)
     out = caplog.text
-    assert (event_detected == account.account_id and
+    assert (event_detected == account_id.primary_identifier and
             ("Could not obtain cached" in out and "No account" not in out))
 
 
 def test_multi_factor_authc_raises_invalid_sequence(
-        default_authenticator, valid_thedude_totp_token, monkeypatch):
+        default_authenticator, valid_thedude_totp_token):
     da = default_authenticator
 
     with pytest.raises(InvalidAuthenticationSequenceException):
@@ -38,19 +39,28 @@ def test_multi_factor_authc_using_accountstore_success(
         caplog, default_authenticator, valid_username_password_token,
         valid_totp_token, event_bus):
     da = default_authenticator
-    event_detected = None
+    success_event_detected = None
+    progress_event_detected = None
 
-    def event_listener(identifiers=None):
-        nonlocal event_detected
-        event_detected = identifiers
-    event_bus.register(event_listener, 'AUTHENTICATION.SUCCEEDED')
+    def progress_event_listener(identifier=None):
+        nonlocal progress_event_detected
+        progress_event_detected = identifier
+    event_bus.register(progress_event_listener, 'AUTHENTICATION.PROGRESS')
 
-    account = da.authenticate_account(None, valid_username_password_token)
+    def success_event_listener(identifier=None):
+        nonlocal success_event_detected
+        success_event_detected = identifier
+    event_bus.register(success_event_listener, 'AUTHENTICATION.SUCCEEDED')
+
+    try:
+        account_id = da.authenticate_account(None, valid_username_password_token)
+    except AdditionalAuthenticationRequired as exc:
+        account_id = da.authenticate_account(exc.account_id, valid_totp_token)
+
     out = caplog.text
-    assert (event_detected == account.account_id and
+    assert (success_event_detected == account_id.primary_identifier and
+            progress_event_detected == account_id.primary_identifier and
             ("Could not obtain cached" in out and "No account" not in out))
-
-
 
 def test_single_factor_authc_userpass_using_cache_success(
         caplog, default_authenticator, invalid_username_password_token,
@@ -107,7 +117,7 @@ def test_single_factor_authc_userpass_using_cache_failure(
         event_detected = username
     event_bus.register(event_listener, 'AUTHENTICATION.FAILED')
 
-    cred = Credential('letsgobowlingggggg')
+    cred = 'letsgobowlingggggg'
     cache_handler.set(domain='credentials', identifier='thedude', value=cred)
 
     with pytest.raises(AuthenticationException):